# cmd/server/handlers.go
package main

import (
	"bufio"
	"encoding/base64"
	"encoding/json"
	"fmt"
	"io"
	"log"
	"mime"
	"net/http"
	"os"
	"strings"
	"time"

	"github.com/b-open-io/bsocial-overlay/bap"
	"github.com/b-open-io/bsocial-overlay/bsocial"
	"github.com/b-open-io/overlay/beef"
	"github.com/bsv-blockchain/go-overlay-services/pkg/core/engine"
	"github.com/bsv-blockchain/go-sdk/overlay"
	"github.com/bsv-blockchain/go-sdk/transaction"
	"github.com/gofiber/fiber/v2"
	"github.com/redis/go-redis/v9"
	"github.com/swaggo/swag/v2"
	scalar "github.com/bdpiprava/scalar-go"
)

// Handlers holds all HTTP handler dependencies
type Handlers struct {
	bapLookup     *bap.LookupService
	bsocialLookup *bsocial.LookupService
	beefStore     *beef.RedisBeefStorage
	engine        *engine.Engine
	cache         *redis.Client
	currentBlock  *func() uint32
	port          int
	subscribe     chan *subRequest
	unsubscribe   chan *subRequest
}

// NewHandlers creates a new Handlers instance with dependencies
func NewHandlers(
	bapLookup *bap.LookupService,
	bsocialLookup *bsocial.LookupService,
	beefStore *beef.RedisBeefStorage,
	eng *engine.Engine,
	cache *redis.Client,
	currentBlock *func() uint32,
	port int,
	subscribe chan *subRequest,
	unsubscribe chan *subRequest,
) *Handlers {
	return &Handlers{
		bapLookup:     bapLookup,
		bsocialLookup: bsocialLookup,
		beefStore:     beefStore,
		engine:        eng,
		cache:         cache,
		currentBlock:  currentBlock,
		port:          port,
		subscribe:     subscribe,
		unsubscribe:   unsubscribe,
	}
}

// @Summary Ingest raw transaction
// @Description Submit a raw BSV transaction for processing. Transaction is validated, input transactions loaded from BEEF store, converted to BEEF format, and submitted to overlay engine for BAP/BSocial indexing.
// @Tags Transaction
// @Accept application/octet-stream
// @Produce json
// @Param transaction body string true "Raw transaction bytes"
// @Success 200 {object} Response{result=object{txid=string}} "Transaction accepted with TXID"
// @Failure 400 {object} Response "Invalid transaction data"
// @Failure 404 {object} Response "Input transaction not found in BEEF store"
// @Failure 500 {object} Response "Failed to process transaction"
// @Router /ingest [post]
func (h *Handlers) IngestTransaction(c *fiber.Ctx) error {
	tx, err := transaction.NewTransactionFromBytes(c.Body())
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(Response{
			Status:  "ERROR",
			Message: "Invalid transaction data: " + err.Error(),
		})
	}

	for _, input := range tx.Inputs {
		sourceBeef, err := h.beefStore.LoadBeef(c.Context(), input.SourceTXID)
		if err != nil {
			return c.Status(fiber.StatusNotFound).JSON(Response{
				Status:  "ERROR",
				Message: "Failed to load input transaction: " + err.Error(),
			})
		}
		input.SourceTransaction, err = transaction.NewTransactionFromBEEF(sourceBeef)
		if err != nil {
			return c.Status(fiber.StatusInternalServerError).JSON(Response{
				Status:  "ERROR",
				Message: "Failed to parse input transaction: " + err.Error(),
			})
		}
	}

	taggedBeef := overlay.TaggedBEEF{
		Topics: []string{"tm_bap", "tm_bsocial"},
	}

	taggedBeef.Beef, err = tx.AtomicBEEF(false)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to generate BEEF: " + err.Error(),
		})
	}

	_, err = h.engine.Submit(c.Context(), taggedBeef, engine.SubmitModeHistorical, nil)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to submit transaction: " + err.Error(),
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: map[string]any{
			"txid": tx.TxID().String(),
		},
	})
}

// @Summary Autofill search
// @Description Fast combined search across identities (top 3) and posts (top 10) for autocomplete/autofill functionality. Results cached for 15 minutes.
// @Tags Search
// @Accept json
// @Produce json
// @Param q query string true "Search query"
// @Success 200 {object} Response{result=object{identities=[]bap.Identity,posts=[]object}} "Combined search results"
// @Failure 500 {object} Response "Search failed"
// @Router /autofill [get]
func (h *Handlers) Autofill(c *fiber.Ctx) error {
	q := c.Query("q")

	if h.cache != nil {
		cached, err := h.cache.HGet(c.Context(), "autofill", q).Result()
		if err != nil {
			log.Println("Cache error:", err)
		} else if cached != "" {
			return c.JSON(Response{
				Status: "OK",
				Result: json.RawMessage(cached),
			})
		}
	}

	identities, err := h.bapLookup.Search(c.Context(), q, 3, 0)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to search identities: " + err.Error(),
		})
	}

	posts, err := h.bsocialLookup.Search(c.Context(), q, 10, 0)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to search posts: " + err.Error(),
		})
	}

	result, err := json.Marshal(map[string]any{
		"identities": identities,
		"posts":      posts,
	})
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to marshal result: " + err.Error(),
		})
	}

	if h.cache != nil {
		h.cache.Pipelined(c.Context(), func(pipe redis.Pipeliner) error {
			pipe.HSet(c.Context(), "autofill", q, result)
			pipe.HExpire(c.Context(), "autofill", 15*time.Minute, q)
			return nil
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: json.RawMessage(result),
	})
}

// @Summary Search identities
// @Description Search for BAP identities using full-text search with MongoDB Atlas Search
// @Tags Identity
// @Accept json
// @Produce json
// @Param q query string true "Search query"
// @Param limit query int false "Results limit" default(20) minimum(1) maximum(100)
// @Param offset query int false "Results offset for pagination" default(0) minimum(0)
// @Success 200 {object} Response{result=[]bap.Identity} "List of matching identities"
// @Failure 500 {object} Response "Internal server error"
// @Router /identity/search [get]
func (h *Handlers) SearchIdentities(c *fiber.Ctx) error {
	q := c.Query("q")
	limit := c.QueryInt("limit", 20)
	offset := c.QueryInt("offset", 0)

	identities, err := h.bapLookup.Search(c.Context(), q, limit, offset)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to search identities: " + err.Error(),
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: identities,
	})
}

// @Summary Search posts
// @Description Search for BSocial posts using full-text search with MongoDB Atlas Search
// @Tags Post
// @Accept json
// @Produce json
// @Param q query string true "Search query"
// @Param limit query int false "Results limit" default(20) minimum(1) maximum(100)
// @Param offset query int false "Results offset for pagination" default(0) minimum(0)
// @Success 200 {object} Response{result=[]object} "List of matching posts with BMAP data"
// @Failure 500 {object} Response "Internal server error"
// @Router /post/search [get]
func (h *Handlers) SearchPosts(c *fiber.Ctx) error {
	q := c.Query("q")
	limit := c.QueryInt("limit", 20)
	offset := c.QueryInt("offset", 0)

	posts, err := h.bsocialLookup.Search(c.Context(), q, limit, offset)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to search posts: " + err.Error(),
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: posts,
	})
}

// @Summary Validate identity by address
// @Description Check if an address was valid for a specific identity at a given block height or timestamp. Used for temporal identity verification with address rotation support.
// @Tags Identity
// @Accept json
// @Produce json
// @Param request body IdentityValidByAddressParams true "Validation parameters (if block and timestamp are 0, uses current block)"
// @Success 200 {object} Response{result=IdentityValidResponse} "Validation result with identity details"
// @Failure 404 {object} Response "Identity not found"
// @Failure 500 {object} Response "Internal server error"
// @Router /identity/validByAddress [post]
func (h *Handlers) ValidateIdentityByAddress(c *fiber.Ctx) error {
	req := &IdentityValidByAddressParams{}
	c.BodyParser(&req)

	if req.Block == 0 && req.Timestamp == 0 {
		req.Block = (*h.currentBlock)()
		// For timestamp, we'd need access to currentBlock.Header.Timestamp
		// but we only have the height function, so leave timestamp at 0
	}

	id, err := h.bapLookup.LoadIdentityByAddress(c.Context(), req.Address)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: err.Error(),
		})
	}

	if id == nil {
		return c.Status(fiber.StatusNotFound).JSON(Response{
			Status:  "ERROR",
			Message: fmt.Sprintf("Identity not found for address %s", req.Address),
		})
	}

	var valid bool
	var currentAddress string

	if req.Block > 0 {
		for _, addr := range id.Addresses {
			if addr.Block <= req.Block {
				currentAddress = addr.Address
			} else {
				break
			}
		}
		valid = currentAddress == req.Address
	} else {
		for _, addr := range id.Addresses {
			if addr.Timestamp <= req.Timestamp {
				currentAddress = addr.Address
			} else {
				break
			}
		}
		valid = currentAddress == req.Address
	}

	result := IdentityValidResponse{
		Identity: *id,
		ValidityRecord: ValidityRecord{
			Valid:     valid,
			Block:     req.Block,
			Timestamp: req.Timestamp,
		},
	}

	if valid {
		result.Profile = id.Profile
	}

	return c.JSON(Response{
		Status: "OK",
		Result: result,
	})
}

// @Summary Get profile image field
// @Description Retrieve and proxy an image from a profile field (e.g., avatar, header). Supports multiple formats: bitfs://, ORDFS URLs, data URLs, HTTP URLs, and raw TxIDs. Returns default avatar if field is empty.
// @Tags Profile
// @Accept json
// @Produce image/jpeg,image/png,image/gif,image/webp
// @Param field path string true "Profile field name (e.g., 'avatar', 'header', 'banner')"
// @Param bapId path string true "BAP Identity ID"
// @Success 200 {file} binary "Image data with appropriate Content-Type header"
// @Failure 400 {object} Response "Invalid field or BAPID"
// @Failure 404 {object} Response "Profile not found"
// @Failure 500 {object} Response "Internal server error or failed to fetch image"
// @Router /person/{field}/{bapId} [get]
func (h *Handlers) GetPersonImageField(c *fiber.Ctx) error {
	field := c.Params("field")
	if field == "" {
		return c.Status(fiber.StatusBadRequest).JSON(Response{
			Status:  "ERROR",
			Message: "Field is required",
		})
	}

	bapId := c.Params("bapId")
	if bapId == "" {
		return c.Status(fiber.StatusBadRequest).JSON(Response{
			Status:  "ERROR",
			Message: "BAPID is required",
		})
	}

	id, err := h.bapLookup.LoadIdentityById(c.Context(), bapId)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: err.Error(),
		})
	}

	if id == nil || id.Profile == nil {
		return c.Status(fiber.StatusNotFound).JSON(Response{
			Status:  "ERROR",
			Message: fmt.Sprintf("Profile not found for BAPID %s", bapId),
		})
	}

	imageUrl, imageExists := id.Profile[field].(string)
	if !imageExists || strings.TrimSpace(imageUrl) == "" {
		imageUrl = "/096b5fdcb6e88f8f0325097acca2784eabd62cd4d1e692946695060aff3d6833_7"
	}

	// Check if the imageUrl is a raw txid (64 character hex string)
	if len(imageUrl) == 64 && !strings.HasPrefix(imageUrl, "/") && !strings.HasPrefix(imageUrl, "http") && !strings.HasPrefix(imageUrl, "data:") {
		imageUrl = "/" + imageUrl
	}

	if strings.HasPrefix(imageUrl, "data:") {
		// Handle base64-encoded data URL
		commaIndex := strings.Index(imageUrl, ",")
		if commaIndex < 0 {
			return c.Status(fiber.StatusBadRequest).JSON(Response{
				Status:  "ERROR",
				Message: "Invalid data URL format",
			})
		}

		metaData := strings.TrimPrefix(imageUrl[:commaIndex], "data:")
		metaDataParts := strings.Split(metaData, ";")
		metaData = metaDataParts[0]
		base64Data := imageUrl[commaIndex+1:]

		mediaType, _, err := mime.ParseMediaType(metaData)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(Response{
				Status:  "ERROR",
				Message: "Invalid media type in data URL " + metaData + " " + err.Error(),
			})
		}

		imgData, err := base64.StdEncoding.DecodeString(base64Data)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(Response{
				Status:  "ERROR",
				Message: "Failed to decode base64 image data",
			})
		}

		c.Set("Content-Type", mediaType)
		return c.Send(imgData)
	}

	// Handle regular image URL
	if strings.HasPrefix(imageUrl, "bitfs://") {
		baseUrl := "https://ordfs.network/"
		path := strings.TrimPrefix(imageUrl, "bitfs://")
		parts := strings.Split(path, ".")
		if len(parts) >= 3 && parts[1] == "out" {
			txid := parts[0]
			imageUrl = baseUrl + txid
		} else {
			return c.Status(fiber.StatusBadRequest).JSON(Response{
				Status:  "ERROR",
				Message: "Invalid bitfs URL format",
			})
		}
	}

	if strings.HasPrefix(imageUrl, "/") {
		imageUrl = "https://ordfs.network" + imageUrl
	}

	resp, err := http.Get(imageUrl)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to fetch image at " + imageUrl + err.Error(),
		})
	}
	defer resp.Body.Close()

	if resp.StatusCode != http.StatusOK {
		return c.Status(fiber.StatusNotFound).JSON(Response{
			Status:  "ERROR",
			Message: "Image not found at the specified URL",
		})
	}

	imgData, err := io.ReadAll(resp.Body)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to read image data",
		})
	}

	contentType := resp.Header.Get("Content-Type")
	if contentType == "" {
		contentType = http.DetectContentType(imgData)
	}

	c.Set("Content-Type", contentType)
	return c.Send(imgData)
}

// @Summary List profiles
// @Description Get a paginated list of BAP profiles with their identity data
// @Tags Profile
// @Accept json
// @Produce json
// @Param limit query int false "Results limit" default(20) minimum(1) maximum(100)
// @Param offset query int false "Results offset for pagination" default(0) minimum(0)
// @Success 200 {object} Response{result=[]bap.Profile} "List of profiles"
// @Failure 500 {object} Response "Internal server error"
// @Router /profile [get]
func (h *Handlers) ListProfiles(c *fiber.Ctx) error {
	offset := c.QueryInt("offset", 0)
	limit := c.QueryInt("limit", 20)

	profiles, err := h.bapLookup.LoadProfiles(c.Context(), limit, offset)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to fetch profiles: " + err.Error(),
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: profiles,
	})
}

// @Summary Get profile by BAP ID
// @Description Retrieve profile data for a specific BAP identity
// @Tags Profile
// @Accept json
// @Produce json
// @Param bapId path string true "BAP Identity ID"
// @Success 200 {object} Response{result=object} "Profile data as key-value map"
// @Failure 500 {object} Response "Internal server error"
// @Router /profile/{bapId} [get]
func (h *Handlers) GetProfileByBapId(c *fiber.Ctx) error {
	bapId := c.Params("bapId")

	identity, err := h.bapLookup.LoadIdentityById(c.Context(), bapId)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to fetch profile: " + err.Error(),
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: identity.Profile,
	})
}

// @Summary Get identity by ID
// @Description Retrieve a complete BAP identity by its ID key (BapId)
// @Tags Identity
// @Accept json
// @Produce json
// @Param request body object{idKey=string} true "Identity request with idKey field"
// @Success 200 {object} Response{result=bap.Identity} "Identity details including addresses, profile, and attestations"
// @Failure 400 {object} Response "Invalid request body"
// @Failure 404 {object} Response "Identity not found"
// @Failure 500 {object} Response "Internal server error"
// @Router /identity/get [post]
func (h *Handlers) GetIdentity(c *fiber.Ctx) error {
	req := map[string]string{}
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(Response{
			Status:  "ERROR",
			Message: "Invalid request body: " + err.Error(),
		})
	}

	id, err := h.bapLookup.LoadIdentityById(c.Context(), req["idKey"])
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to fetch identity: " + err.Error(),
		})
	}

	if id == nil {
		return c.Status(fiber.StatusNotFound).JSON(Response{
			Status:  "ERROR",
			Message: "Identity not found for ID key: " + req["idKey"],
		})
	}

	return c.JSON(Response{
		Status: "OK",
		Result: id,
	})
}

// @Summary Subscribe to real-time updates
// @Description Subscribe to real-time overlay updates via Server-Sent Events (SSE). Streams updates from Redis PubSub channels for specified topics. Connection is long-lived and automatically cleaned up on disconnect.
// @Tags Subscription
// @Produce text/event-stream
// @Param topics path string true "Comma-separated topic list (e.g., 'tm_bap,tm_bsocial')"
// @Success 200 {string} string "Event stream with SSE format: id (timestamp), event (channel), data (payload)"
// @Failure 400 {object} object{error=string} "Missing topics parameter"
// @Failure 404 {object} object{error=string} "No topics provided"
// @Header 200 {string} Content-Type "text/event-stream"
// @Header 200 {string} Cache-Control "no-cache"
// @Header 200 {string} Connection "keep-alive"
// @Router /subscribe/{topics} [get]
func (h *Handlers) SubscribeToTopics(c *fiber.Ctx) error {
	topicsParam := c.Params("topics")
	if topicsParam == "" {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "Missing topics",
		})
	}

	topics := strings.Split(topicsParam, ",")
	if len(topics) == 0 {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": "No topics provided",
		})
	}

	// Set headers for SSE
	c.Set("Content-Type", "text/event-stream")
	c.Set("Cache-Control", "no-cache")
	c.Set("Connection", "keep-alive")
	c.Set("Transfer-Encoding", "chunked")

	subReq := &subRequest{
		topics:  topics,
		msgChan: make(chan *redis.Message, 25),
	}
	h.subscribe <- subReq

	disconnected := make(chan struct{})

	ctx := c.Context()
	ctx.SetBodyStreamWriter(func(w *bufio.Writer) {
		fmt.Println("Client connected for topics:", topics)

		go func() {
			<-ctx.Done()
			fmt.Println("Context done, closing connection for topics:", topics)
			h.unsubscribe <- subReq
			close(disconnected)
		}()

		for {
			select {
			case <-disconnected:
				return
			case msg := <-subReq.msgChan:
				fmt.Fprintf(w, "id: %d\n", time.Now().UnixNano())
				fmt.Fprintf(w, "event: %s\n", msg.Channel)
				fmt.Fprintf(w, "data: %s\n\n", msg.Payload)

				if err := w.Flush(); err != nil {
					fmt.Printf("Error while flushing: %v. Closing connection.\n", err)
					h.unsubscribe <- subReq
					return
				}
			}
		}
	})

	return nil
}

// @Summary Serve OpenAPI specification
// @Description Returns the OpenAPI 3.0 specification in JSON format
// @Tags Documentation
// @Produce json
// @Success 200 {object} object "OpenAPI 3.0 specification"
// @Router /openapi.json [get]
func (h *Handlers) ServeOpenAPISpec(c *fiber.Ctx) error {
	spec, err := swag.ReadDoc("swagger")
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to read OpenAPI spec: " + err.Error(),
		})
	}
	c.Set("Content-Type", "application/json")
	return c.SendString(spec)
}

// @Summary API Documentation UI
// @Description Interactive API documentation powered by Scalar
// @Tags Documentation
// @Produce text/html
// @Success 200 {string} string "HTML documentation page"
// @Router /docs [get]
func (h *Handlers) ServeDocumentation(c *fiber.Ctx) error {
	hostingURL := os.Getenv("HOSTING_URL")
	if hostingURL == "" {
		hostingURL = fmt.Sprintf("http://localhost:%d", h.port)
	}

	html, err := scalar.NewV2(
		scalar.WithBaseServerURL(hostingURL+"/openapi.json"),
		scalar.WithDarkMode(),
	)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(Response{
			Status:  "ERROR",
			Message: "Failed to render documentation: " + err.Error(),
		})
	}

	c.Set("Content-Type", "text/html; charset=utf-8")
	return c.SendString(html)
}